+++ ShoppingCard/Models/CartItemModel.cs
@@ -3,7 +3,7 @@ namespace ShoppingCard.Models
 {
     public class CartItemModel
     {
-        public int ProductId { get; set; } // Nhoi quay lai check tu bai 42
+        public long ProductId { get; set; } // Nhoi quay lai check tu bai 42
 
         public string ProductName { get; set; }
         public int Quantity { get; set; }
